Use a uniqe version number.
To avoid shadowing a Hackage version. Empirically a Cabal sandbox uses
our local version instead of the Hackage version when the version
numbers are the same, but I can't find this documented anywhere. I'm
changing the version number now because Ken and Michael and I just
solved a confusing bug whereby a new version of `base-compat` on
Hackage, 0.9.1 released on April 11, was taking precedence over our
local 0.9.0, and breaking the build due to the `unix` dependency
coming back. This was especially confusing because Michael and I had
Hackage databases which had not been updated since before April 11,
and so for us the local 0.9.0 version was being used, instead of the
Hackage 0.9.0 version.