Match Data.Tuple API in base-4.18
In `base-4.18`, `Data.Tuple` exports both the `MkSolo` data constructor as well
as the `Solo` pattern synonym, the latter being offered for backwards
compatibility. This matches the changes on the `base-compat` and
`base-compat-batteries` side.  Note that `base-compat-batteries` now requires
`OneTuple-0.4` or later on GHC 7.4+, as that is the first `OneTuple` release to
offer `MkSolo`.