base-compat-batteries-0.12.*: Allow building with OneTuple-0.4.*
We make an effort to ensure some degree of API consistency regardless of which
GHC version is used to build the library. In particular, the
`base-compat-batteries-0.12.*` has implicitly chose the convention of using
`Solo` as the name of the data constructor, so we attempt to preserve this
convention. We make an exception for GHC 9.6 (where both `MkSolo` and `Solo`
are available).

Fixes #91.