Make type-check-test work with GHC 9.10
This requires using a more recent version of `typediff` (0.1.6) that is aware
of the module names in `ghc-internal`. It also requires incurring a dependency
on `ghc-internal` itself to ensure that GHCi prints things like
`GHC.Internal.Maybe.Maybe` rather than `ghc-internal:GHC.Internal.Maybe.Maybe`,
the latter of which `typediff` cannot parse.