Drop support for pre-8.0 versions of GHC
Unfortunately, it is no longer possible to reasonably test versions of GHC
prior to 8.0 on modern versions of Linux, macOS, or Windows, as these versions
of GHC all run into subtle OS incompatibility issues. While we could leave all
of the old, pre-8.0 code paths around in `base-compat`, this runs the risk of
bitrotting very quickly. As such, I have opted to remove these code paths here.
Given that GHC 7.10.3 was released nearly a decade ago, I believe that the risk
of breaking anyone's workflows by doing this is very minimal.

(It's somewhat unfortunate that we have to drop compatibility like this in an
explicitly compatibility-oriented library like `base-compat`, but so be it.)